fix: add missing enums to the exposed model, rename data.py to model.py
BREAKING CHANGE: `kosmorrolib.data` has been renamed to
`kosmorrolib.model`. To ensure further BC-break to happen on this side,
prefer using the model now exposed from kosmorrolib directly.
