+++ kosmorrolib/__init__.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
-from .data import Position
+from .model import Position, Event, AsterEphemerides, Object
 from .ephemerides import get_ephemerides, get_moon_phase
 from .events import get_events
+from .enum import EventType, MoonPhaseType
